requirements.txt
tavily-python
langchain-openai
python-dotenv
pandas 